requirements.txt
flask>=3.0
flask-cors>=4.0
flask-jwt-extended>=4.6
pyyaml>=6.0
scapy>=2.5
psutil>=5.9
requests>=2.31
watchdog>=3.0
colorama>=0.4
python-dateutil>=2.8
